+++ components/layout/layout_thread.rs
@@ -34,7 +34,6 @@ use ipc_channel::router::ROUTER;
 use layout_debug;
 use layout_traits::LayoutThreadFactory;
 use log;
-use msg::ParseErrorReporter;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId};
 use net_traits::image_cache_thread::{ImageCacheChan, ImageCacheResult, ImageCacheThread};
 use parallel;
@@ -65,6 +64,7 @@ use style::animation::Animation;
 use style::computed_values::{filter, mix_blend_mode};
 use style::context::{SharedStyleContext, StylistWrapper, ReflowGoal};
 use style::dom::{TDocument, TElement, TNode};
+use style::error_reporting::ParseErrorReporter;
 use style::media_queries::{Device, MediaType};
 use style::parallel::WorkQueueData;
 use style::selector_matching::{Stylist, USER_OR_USER_AGENT_STYLESHEETS};
