+++ components/msg/lib.rs
@@ -7,7 +7,6 @@
 
 #[macro_use]
 extern crate bitflags;
-extern crate cssparser;
 extern crate euclid;
 extern crate hyper;
 extern crate ipc_channel;
@@ -20,12 +19,3 @@ extern crate util;
 pub mod compositor_msg;
 pub mod constellation_msg;
 pub mod webdriver_msg;
-
-use constellation_msg::PipelineId;
-use cssparser::{Parser, SourcePosition};
-
-pub trait ParseErrorReporter {
-    fn report_error(&self, input: &mut Parser, position: SourcePosition, message: &str);
-    fn clone(&self) -> Box<ParseErrorReporter + Send + Sync>;
-    fn pipeline(&self) -> PipelineId;
-}
