+++ components/script/dom/element.rs
@@ -65,7 +65,6 @@ use html5ever::serialize::SerializeOpts;
 use html5ever::serialize::TraversalScope;
 use html5ever::serialize::TraversalScope::{ChildrenOnly, IncludeNode};
 use html5ever::tree_builder::{LimitedQuirks, NoQuirks, Quirks};
-use msg::ParseErrorReporter;
 use selectors::matching::{DeclarationBlock, matches};
 use selectors::matching::{common_style_affecting_attributes, rare_style_affecting_attributes};
 use selectors::parser::{AttrSelector, NamespaceConstraint, parse_author_origin_selector_list_from_str};
@@ -78,6 +77,7 @@ use std::default::Default;
 use std::mem;
 use std::sync::Arc;
 use string_cache::{Atom, Namespace, QualName};
+use style::error_reporting::ParseErrorReporter;
 use style::properties::DeclaredValue;
 use style::properties::longhands::{self, background_image, border_spacing, font_family, font_size};
 use style::properties::{PropertyDeclaration, PropertyDeclarationBlock, parse_style_attribute};
