+++ components/style/context.rs
@@ -5,9 +5,9 @@
 use animation::Animation;
 use app_units::Au;
 use dom::OpaqueNode;
+use error_reporting::ParseErrorReporter;
 use euclid::Size2D;
 use matching::{ApplicableDeclarationsCache, StyleSharingCandidateCache};
-use msg::ParseErrorReporter;
 use selector_matching::Stylist;
 use std::cell::RefCell;
 use std::collections::HashMap;
