+++ tests/unit/style/viewport.rs
@@ -6,7 +6,7 @@ use cssparser::Parser;
 use euclid::scale_factor::ScaleFactor;
 use euclid::size::Size2D;
 use media_queries::CSSErrorReporterTest;
-use msg::ParseErrorReporter;
+use style::error_reporting::ParseErrorReporter;
 use style::media_queries::{Device, MediaType};
 use style::parser::ParserContext;
 use style::stylesheets::{Origin, Stylesheet, CSSRuleIteratorExt};
